changed how reference sequence is processed via a motif and confirmed training works with methylated bases
