+++ src/signalalign/tests/test_trainModels.py
@@ -69,12 +69,18 @@ class TrainSignalAlignTest(unittest.TestCase):
             "singleLevelFixedCanonical": 14
         }
         cls.test_hdp_training_data = os.path.join(cls.HOME, "tests/test_hdp/test_hdp_alignment.txt")
-
+        cls.one_file_dir = os.path.join(cls.HOME, "tests/minion_test_reads/one_R9_canonical_ecoli")
         cls.config_file = os.path.join(cls.HOME, "tests/trainModels-config.json")
         cls.default_args = create_dot_dict(load_json(cls.config_file))
         cls.default_args.path_to_bin = cls.path_to_bin
+        cls.default_args.output_dir = cls.path_to_bin
+        cls.default_args.samples[0].fast5_dirs = [cls.one_file_dir]
+        cls.default_args.samples[0].bwa_reference = cls.ecoli_reference
+        cls.r9_complement_model_file_acgt = os.path.join(cls.HOME, "models/testModelR9_5mer_acgt_complement.model")
+        cls.r9_template_model_file_acgt = os.path.join(cls.HOME, "models/testModelR9_5mer_acgt_template.model")
 
-        cls.one_file_dir = os.path.join(cls.HOME, "tests/minion_test_reads/one_R9_canonical_ecoli")
+        cls.default_args.complement_hmm_model = cls.r9_complement_model_file_acgt
+        cls.default_args.template_hmm_model = cls.r9_template_model_file_acgt
 
     def test_parse_assignment_file(self):
         assignments = parse_assignment_file(self.assignment_file)
@@ -311,7 +317,6 @@ class TrainSignalAlignTest(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tempdir:
             fake_args = create_dot_dict(self.default_args.copy())
             fake_args.output_dir = tempdir
-            fake_args.samples[0].fast5_dirs = [self.one_file_dir]
             fake_args.job_count = 1
             fake_args.training.transitions = True
             fake_args.training.normal_emissions = True
@@ -328,7 +333,6 @@ class TrainSignalAlignTest(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tempdir:
             fake_args = create_dot_dict(self.default_args.copy())
             fake_args.output_dir = tempdir
-            fake_args.samples[0].fast5_dirs = [self.one_file_dir]
             fake_args.training.transitions = True
             fake_args.training.normal_emissions = True
             fake_args.training.hdp_emissions = True
@@ -344,8 +348,6 @@ class TrainSignalAlignTest(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tempdir:
             fake_args = create_dot_dict(self.default_args.copy())
             fake_args.output_dir = tempdir
-            fake_args.samples[0].fast5_dirs = [self.one_file_dir]
-            fake_args.path_to_bin = self.path_to_bin
             fake_args.training.transitions = True
             fake_args.training.normal_emissions = True
             fake_args.training.hdp_emissions = False
@@ -358,7 +360,6 @@ class TrainSignalAlignTest(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tempdir:
             fake_args = create_dot_dict(self.default_args.copy())
             fake_args.output_dir = tempdir
-            fake_args.samples[0].fast5_dirs = [self.one_file_dir]
             fake_args.training.transitions = False
             fake_args.training.normal_emissions = True
             fake_args.training.hdp_emissions = True
@@ -373,6 +374,52 @@ class TrainSignalAlignTest(unittest.TestCase):
             fake_args.training.hdp_emissions = False
             # raise error when no training is selected
             self.assertRaises(AssertionError, TrainSignalAlign(fake_args).expectation_maximization_training)
+
+    def test_methylated_transitions_training(self):
+        with tempfile.TemporaryDirectory() as tempdir:
+            fake_args = create_dot_dict(self.default_args.copy())
+            fake_args.output_dir = tempdir
+            fake_args.samples.append(create_dot_dict(fake_args.samples[0]))
+            fake_args.samples[1].motifs = [["CCAGG", "CEAGG"], ["CCTGG", "CETGG"]]
+            fake_args.samples[1].name = "methylated"
+            fake_args.complement_hmm_model = self.r9_complement_model_file
+            fake_args.template_hmm_model = self.r9_template_model_file
+
+            fake_args.job_count = 1
+            fake_args.training.transitions = True
+            fake_args.training.normal_emissions = False
+            fake_args.training.hdp_emissions = False
+            fake_args.training.expectation_maximization = False
+            # Test EM training 3 rounds
+            template_hmm_model_path, complement_hmm_model_path, template_hdp_model_path, complement_hdp_model_path = \
+                TrainSignalAlign(fake_args).expectation_maximization_training()
+
+    def test_methylated_hdp_training(self):
+        with tempfile.TemporaryDirectory() as tempdir:
+            fake_args = create_dot_dict(self.default_args.copy())
+            fake_args.output_dir = tempdir
+            fake_args.samples.append(create_dot_dict(fake_args.samples[0]))
+            fake_args.samples[1].motifs = [["CCAGG", "CEAGG"], ["CCTGG", "CETGG"]]
+            fake_args.samples[1].name = "methylated"
+            fake_args.complement_hmm_model = self.r9_complement_model_file
+            fake_args.template_hmm_model = self.r9_template_model_file
+
+            fake_args.job_count = 1
+            fake_args.training.transitions = False
+            fake_args.training.normal_emissions = False
+            fake_args.training.hdp_emissions = True
+            fake_args.training.expectation_maximization = False
+            fake_args.training.em_iterations = 3
+            fake_args.hdp_args.hdp_type = "singleLevelPrior2"
+            fake_args.hdp_args.gibbs_samples = 100
+            fake_args.hdp_args.burnin_multiplier = 0
+            fake_args.hdp_args.number_of_assignments = 1
+            fake_args.two_d = True
+
+            # Test EM training 3 rounds
+            template_hmm_model_path, complement_hmm_model_path, template_hdp_model_path, complement_hdp_model_path = \
+                TrainSignalAlign(fake_args).expectation_maximization_training()
+
 
 
 if __name__ == '__main__':
